# fix: moved js filter to mongoDB