--- a/app/server/appsmith-server/src/main/java/com/appsmith/server/newactions/base/NewActionServiceCEImpl.java
+++ b/app/server/appsmith-server/src/main/java/com/appsmith/server/newactions/base/NewActionServiceCEImpl.java
@@ -734,10 +734,10 @@ public class NewActionServiceCEImpl extends BaseService<NewActionRepository, New
     }
 
     @Override
-    public Flux<NewAction> findAllByApplicationIdAndViewMode(
-            String applicationId, Boolean viewMode, AclPermission permission, Sort sort) {
+    public Flux<NewAction> findAllByApplicationIdAndPluginType(
+            String applicationId, Boolean viewMode, AclPermission permission, Sort sort, List<String> pluginTypes) {
         return repository
-                .findByApplicationId(applicationId, permission, sort)
+                .findByApplicationIdAndPluginType(applicationId, pluginTypes, permission, sort)
                 .name(VIEW_MODE_FETCH_ACTIONS_FROM_DB)
                 .tap(Micrometer.observation(observationRegistry))
                 // In case of view mode being true, filter out all the actions which haven't been published
@@ -759,6 +759,28 @@ public class NewActionServiceCEImpl extends BaseService<NewActionRepository, New
                 .flatMap(this::sanitizeAction)
                 .name(VIEW_MODE_SANITISE_ACTION)
                 .tap(Micrometer.observation(observationRegistry));
+    }
+
+    @Override
+    public Flux<NewAction> findAllByApplicationIdAndViewMode(
+            String applicationId, Boolean viewMode, AclPermission permission, Sort sort) {
+        return repository
+                .findByApplicationId(applicationId, permission, sort)
+                // In case of view mode being true, filter out all the actions which haven't been published
+                .flatMap(action -> {
+                    if (Boolean.TRUE.equals(viewMode)) {
+                        // In case we are trying to fetch published actions but this action has not been published, do
+                        // not return
+                        if (action.getPublishedAction() == null) {
+                            return Mono.empty();
+                        }
+                    }
+                    // No need to handle the edge case of unpublished action not being present. This is not possible
+                    // because every created action starts from an unpublishedAction state.
+
+                    return Mono.just(action);
+                })
+                .flatMap(this::sanitizeAction);
     }
 
     @Override
@@ -792,9 +814,18 @@ public class NewActionServiceCEImpl extends BaseService<NewActionRepository, New
             return Flux.error(new AppsmithException(AppsmithError.INVALID_PARAMETER, FieldName.APPLICATION_ID));
         }
 
+        List<String> pluginTypes = List.of(
+                PluginType.DB.toString(),
+                PluginType.API.toString(),
+                PluginType.SAAS.toString(),
+                PluginType.REMOTE.toString(),
+                PluginType.AI.toString(),
+                PluginType.INTERNAL.toString());
+
         // fetch the published actions by applicationId
         // No need to sort the results
-        return findAllByApplicationIdAndViewMode(applicationId, true, actionPermission.getExecutePermission(), null)
+        return findAllByApplicationIdAndPluginType(
+                        applicationId, true, actionPermission.getExecutePermission(), null, pluginTypes)
                 .name(VIEW_MODE_INITIAL_ACTION)
                 .tap(Micrometer.observation(observationRegistry))
                 .filter(newAction -> !PluginType.JS.equals(newAction.getPluginType()))