@sneha122 2024-08-22
> @sneha122 Could we please add or modify any test cases relevant to the new logic we’ve implemented? Additionally, please confirm if no new test cases are necessary in this context.

@sagar-qa007 This wont require any test cases as the functioning would be same only performance would be improved.